railway-barcode-system-new/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
qrcode[pil]==7.4.2
Pillow==10.3.0
pyzbar==0.1.9
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
requests==2.31.0
PyJWT==2.8.0
pytesseract==0.3.10
scipy==1.11.4
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196

# Optional heavy ML (uncomment if needed)
# tensorflow-cpu==2.12.0
